pyproject.toml
[project]
name = "lite-bootstrap"
description = "Lite package for bootstrapping new microservices"
authors = [
    { name = "Artur Shiriev", email = "me@shiriev.ru" },
]
readme = "README.md"
requires-python = ">=3.10,<4"
license = "MIT"
keywords = [
    "python",
    "microservice",
    "bootstrap",
    "opentelemetry",
    "sentry",
    "error-tracing",
    "fastapi",
    "structlog",
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Software Development :: Libraries",
]
dynamic = ["version"]
packages = [
    { include = "lite_bootstrap" },
]

[project.urls]
repository = "https://github.com/modern-python/lite-bootstrap"

[project.optional-dependencies]
sentry = [
    "sentry-sdk",
]
otl = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation",
]
logging = [
    "structlog",
]
fastapi = [
    "fastapi",
]
fastapi-otl = [
    "opentelemetry-instrumentation-fastapi",
]
fastapi-all = [
    "lite-bootstrap[sentry,otl,logging,fastapi,fastapi-otl]"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "httpx",  # for test client
    "mypy",
    "ruff",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1", # allow missing docstrings
    "S101", # allow asserts
    "TCH", # ignore flake8-type-checking
    "FBT", # allow boolean args
    "D203", # "one-blank-line-before-class" conflicting with D211
    "D213", # "multi-line-summary-second-line" conflicting with D212
    "COM812", # flake8-commas "Trailing comma missing"
    "ISC001", # flake8-implicit-str-concat
]
isort.lines-after-imports = 2
isort.no-lines-before = ["standard-library", "local-folder"]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report term-missing"

[tool.coverage.report]
exclude_also = ["if typing.TYPE_CHECKING:"]
